requirements.txt
streamlit
langchain_groq
langchain_community
pypdf
sentence-transformers
pinecone